+++ src/internal/cpu/cpu_arm64_android.go
@@ -2,6 +2,10 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build arm64
+
 package cpu
 
-const GOOS = "android"
+func osInit() {
+	hwcapInit("android")
+}
