+++ src/internal/cpu/cpu_arm64_linux.go
@@ -2,10 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build !linux
-// +build !freebsd
+// +build arm64
+// +build linux
 // +build !android
 
 package cpu
 
-const GOOS = "other"
+func osInit() {
+	hwcapInit("linux")
+}
