--- src/internal/cpu/cpu_linux.go
@@ -1,9 +1,0 @@
-// Copyright 2020 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// +build !android
-
-package cpu
-
-const GOOS = "linux"
